Support poll operation in inode, only implemented for dev and pipe.
